UI아이콘작업중

+++ CodeVein_0.1/Client/Codes/Expendables_Inven.cpp
@@ -30,10 +30,10 @@ HRESULT CExpendables_Inven::Ready_GameObject(void * pArg)
 
 	CUI::Ready_GameObject(pArg);
 
-	m_fPosX = WINCX * 0.75f;
+	m_fPosX = WINCX * 0.3f;
 	m_fPosY = WINCY * 0.5f;
-	m_fSizeX = WINCX * 0.5f;
-	m_fSizeY = WINCY;
+	m_fSizeX = 280.f;
+	m_fSizeY = 471.f;
 
 	m_fViewZ = 1.f;
 
@@ -42,25 +42,28 @@ HRESULT CExpendables_Inven::Ready_GameObject(void * pArg)
 	// Slot Create
 	CUI::UI_DESC* pDesc = nullptr;
 	CExpendables_Slot* pExSlot = nullptr;
-	for (_uint i = 0; i < 2; ++i)
+	for (_uint i = 0; i < 6; ++i)
 	{
 		for (_uint j = 0; j < 5; ++j)
 		{
 			pDesc = new CUI::UI_DESC;
-			pDesc->fPosX = m_fPosX - 180.f + 90.f * j;
-			pDesc->fPosY = m_fPosY - 100.f + 90.f * i;
-			pDesc->fSizeX = 80.f;
-			pDesc->fSizeY = 80.f;
+			pDesc->fPosX = m_fPosX - 103.f + 52.f * j;
+			pDesc->fPosY = m_fPosY - 130.f + 52.f * i;
+			pDesc->fSizeX = 50.f;
+			pDesc->fSizeY = 50.f;
 			g_pManagement->Add_GameObject_ToLayer(L"GameObject_ExpendSlot", SCENE_STAGE, L"Layer_ExpendSlot", pDesc);
 			pExSlot = static_cast<CExpendables_Slot*>(g_pManagement->Get_GameObjectBack(L"Layer_ExpendSlot", SCENE_STAGE));
 
 			m_vecSlot.push_back(pExSlot);
+			m_vecUI_DESC.push_back(pDesc);
 		}
-
 	}
-	Add_Expendables(CExpendables::EXPEND_1);
-	Add_Expendables(CExpendables::EXPEND_2);
-	Add_Expendables(CExpendables::EXPEND_3);
+
+	Add_MultiExpendables(CExpendables::EXPEND_1, 15);
+	Add_MultiExpendables(CExpendables::EXPEND_2, 15);
+	Add_MultiExpendables(CExpendables::EXPEND_3, 15);
+	Add_MultiExpendables(CExpendables::EXPEND_4, 20);
+
 	return NOERROR;
 }
 
@@ -74,9 +77,8 @@ _int CExpendables_Inven::Update_GameObject(_double TimeDelta)
 	m_pRendererCom->Add_RenderList(RENDER_UI, this);
 
 	D3DXMatrixOrthoLH(&m_matProj, WINCX, WINCY, 0.f, 1.0f);
-
-	if (m_bIsActive)
-		Click_Inven();
+	
+	Click_Inven();
 
 	for (auto& pSlot : m_vecSlot)
 	{
@@ -124,7 +126,7 @@ HRESULT CExpendables_Inven::Render_GameObject()
 
 	m_pShaderCom->Begin_Shader();
 
-	m_pShaderCom->Begin_Pass(0);
+	m_pShaderCom->Begin_Pass(1);
 
 	
 	m_pBufferCom->Render_VIBuffer();
@@ -184,6 +186,8 @@ HRESULT CExpendables_Inven::SetUp_ConstantTable()
 
 void CExpendables_Inven::Click_Inven()
 {
+	if (!m_bIsActive)
+		return;
 
 	for (auto& pExpendSlot : m_vecSlot)
 	{
@@ -203,10 +207,11 @@ void CExpendables_Inven::Click_Inven()
 			}
 		}
 	}
-	
 
-	//if (g_pInput_Device->Key_Up(DIK_6))
-	//	Sell_Expendables(2);
+	/*if (g_pInput_Device->Key_Up(DIK_6))
+		Add_MultiExpendables(CExpendables::EXPEND_1, 1);
+	if (g_pInput_Device->Key_Up(DIK_7))
+		Sell_Expendables(1);*/
 }
 
 void CExpendables_Inven::Add_Expendables(CExpendables::EXPEND_TYPE eType)
@@ -221,6 +226,12 @@ void CExpendables_Inven::Add_Expendables(CExpendables::EXPEND_TYPE eType)
 	pExpendables->Set_Type(eType);
 	
 	Load_Expendables(pExpendables, 0);
+}
+
+void CExpendables_Inven::Add_MultiExpendables(CExpendables::EXPEND_TYPE eType, _uint iCnt)
+{
+	for (_uint i = 0; i < iCnt; ++i)
+		Add_Expendables(eType);
 }
 
 void CExpendables_Inven::Delete_QuickSlot(CExpendables_Slot* pSlot)
@@ -239,14 +250,35 @@ void CExpendables_Inven::Delete_QuickSlot(CExpendables_Slot* pSlot)
 
 void CExpendables_Inven::Sell_Expendables(_uint iDelete)
 {
+	_uint idx = 0;
 	for (auto& pSlot : m_vecSlot)
 	{
 		if (pSlot->Pt_InRect())
 		{
+			if (iDelete > pSlot->Get_Size())
+				return;
 			for (_uint i = 0; i < iDelete; ++i)
+			{
 				pSlot->Delete_Item();
-			return;
+			}
+					
+			if (pSlot->Get_Size() == 0)
+			{
+				m_vecSlot.erase(m_vecSlot.begin() + idx);
+				m_vecSlot.shrink_to_fit();
+				break;
+			}
 		}
+		
+		++idx;
+	}
+
+	_ulong Idx = 0;
+	for (auto& pSlot : m_vecSlot)
+	{
+
+		pSlot->Set_UI_Pos(m_vecUI_DESC[Idx]->fPosX, m_vecUI_DESC[Idx]->fPosY);
+		++Idx;
 	}
 }
 
@@ -255,11 +287,11 @@ void CExpendables_Inven::Load_Expendables(CExpendables* pExpendables, _uint iInd
 {
 	if (nullptr == pExpendables)
 		return;
-	if (m_vecSlot.size() < iIndex)
+	if (m_vecSlot.size() <= iIndex)
 		return;
 
-	if (m_vecSlot[iIndex]->Get_Type() == pExpendables->Get_Type() ||
-		m_vecSlot[iIndex]->Get_Size() == 0)
+	if ((m_vecSlot[iIndex]->Get_Type() == pExpendables->Get_Type() ||
+		m_vecSlot[iIndex]->Get_Size() == 0) && m_vecSlot[iIndex]->Get_Size() < 9)
 		m_vecSlot[iIndex]->Input_Item(pExpendables);
 	else
 		Load_Expendables(pExpendables, iIndex + 1);
